requirements.txt
fastapi==0.115.6
PyMySQL==1.1.1
python-dotenv==1.0.1
